fix(category): fetchCategories after each CRUD operation feat(category): add CategoryBoard component
